refactor(pgwire): split ddl/dsl.rs into per-concern modules
dsl.rs has grown to cover DDL parsing for search indexes, CRDT merge,
FTS, vector, sparse, and fusion in a single flat file. Split it into a
dsl/ directory with one module per concern:

  crdt_merge, fulltext_index, search_fusion, search_index,
  search_vector, sparse_index, vector_index, helpers, mod

No logic changes — pure file reorganisation.
